services/session_provider/requirements.txt
aiokafka
redis
redis[hiredis]   
pytest
python-dotenv
passlib
python-multipart
apscheduler